+++ scripts/developer-tools/build-csv-exports-files-from-old-local-database.sh
@@ -11,4 +11,6 @@ php ./bin/console app:import:artist ./var/csv/imports/artist.csv
 php ./bin/console app:import:menu:level1 ./var/csv/imports/menulevel1.csv
 php ./bin/console app:import:menu:level2 ./var/csv/imports/menulevel2.csv
 php ./bin/console app:import:page ./var/csv/imports/page.csv
+cp -r ../lo-pati/web/uploads/artists/* public/uploads/artists/
+cp ../lo-pati/web/uploads/slides/* public/uploads/slides/
 echo "Finished at $(date +"%T %d/%m/%Y")"