+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 ##### Version 5.02.00 (WIP)
  * add frontend entities
+ * add frontend homepage slideshow
 
 ##### Version 5.01.00 (2021-07-13)
  * add menu rendering